ci(workflows): add volume and working directory to docker run commands
The changes ensure that the Docker container has access to the current working directory by mounting it as a volume and setting it as the working directory. This is necessary for the R script to correctly write the installed packages CSV file and for the subsequent bash script to update the README files.